Create codeBasev1
This is where my code has started. At this point, I can get a command window to open an accept the user input, but I haven't had luck getting the command to run in the background and produce the proper result. The way my code is written at the moment, it opens a new file to run the code but doesn't generate any results.....yet.