Create code v2
This is the first level of the functioning application. It will successfully print the network key for a given SSID as long as the computer it is being run on has successfully connected to the wi-fi network in question. My next step will be to clean up the output to where the only text that is being returned to the user is the network key itself.